random is working

+++ game-ctrl/myrand.h
@@ -2,7 +2,7 @@
 #define _MYRAND_H_
 
 void myRandInit();
-uint8_t myRandGet();
+uint16_t myRandGet();
 
 
 #endif // _MYRAND_H_
